Add HasParticipatoryProcess controller concern
