+++ app/controllers/action_plans_controller.rb
@@ -1,9 +1,9 @@
 class ActionPlansController < ApplicationController
+  include HasParticipatoryProcess
   include ModerateActions
 
   load_and_authorize_resource
   respond_to :html, :json
-  before_action :load_participation_process
 
   def index
   end
