+++ app/controllers/meetings_controller.rb
@@ -1,8 +1,7 @@
 class MeetingsController < ApplicationController
+  include HasParticipatoryProcess
   load_and_authorize_resource
   respond_to :html, :json
-
-  before_action :load_participation_process
 
   def index
   end
